Text Add, Size functionality
Next Idea - Give a button on top right to complete the text change and reflect that on the previous page. Complete also takes back to the previous page. But, before complete performs function it should give a prompt that "Are you sure you want to change the text - Yes/No". Once the this functionality is done, then i'll implement change color on font, its border on the main page. and also change border size on the text page.